Fix MAC file loading; bump patch version
