+++ src/main/resources/project.properties
@@ -1,1 +1,1 @@
-version = 1.2.1
+version = 1.2.12
